Move restart-gather test into gather_test.go
